add modal analisis

+++ database/migrations/2024_03_23_024103_create_isi_form_table.php
@@ -8,13 +8,15 @@ class CreateIsiFormTable extends Migration
 {
     public function up()
     {
-        Schema::create('isi_form', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('formulir_id');
-            $table->foreign('formulir_id')->references('id')->on('formulir')->onDelete('cascade');
-            $table->text('isi');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('isi_form')) {
+            Schema::create('isi_form', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('formulir_id');
+                $table->foreign('formulir_id')->references('id')->on('formulir')->onDelete('cascade');
+                $table->text('isi');
+                $table->timestamps();
+            });
+        }
     }
 
     public function down()
