+++ database/migrations/2024_04_04_132827_create_kelengkapan_table.php
@@ -14,23 +14,25 @@ class CreateKelengkapanTable extends Migration
      */
     public function up()
     {
-        Schema::create('kelengkapan', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('formulir_id');
-            $table->unsignedBigInteger('isi_form_id');
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('pasien_id');
-            $table->enum('kualitatif', ['lengkap', 'tidak lengkap']);
-            $table->enum('kuantitatif', ['lengkap', 'tidak lengkap']);
-            $table->timestamp('tglberkas')->nullable();
-            $table->timestamp('tglcek')->nullable()->default(DB::raw('CURRENT_TIMESTAMP'));
-            $table->timestamps();
+        if (!Schema::hasTable('kelengkapan')) {
+            Schema::create('kelengkapan', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('formulir_id');
+                $table->unsignedBigInteger('isi_form_id');
+                $table->unsignedBigInteger('user_id');
+                $table->unsignedBigInteger('pasien_id');
+                $table->enum('kualitatif', ['lengkap', 'tidak lengkap']);
+                $table->enum('kuantitatif', ['lengkap', 'tidak lengkap']);
+                $table->timestamp('tglberkas')->nullable();
+                $table->timestamp('tglcek')->nullable()->default(DB::raw('CURRENT_TIMESTAMP'));
+                $table->timestamps();
 
-            $table->foreign('formulir_id')->references('id')->on('formulirs')->onDelete('cascade');
-            $table->foreign('isi_form_id')->references('id')->on('isi_forms');
-            $table->foreign('user_id')->references('id')->on('users');
-            $table->foreign('pasien_id')->references('id')->on('pasiens');
-        });
+                $table->foreign('formulir_id')->references('id')->on('formulirs')->onDelete('cascade');
+                $table->foreign('isi_form_id')->references('id')->on('isi_forms');
+                $table->foreign('user_id')->references('id')->on('users');
+                $table->foreign('pasien_id')->references('id')->on('pasiens');
+            });
+        }
     }
 
     /**
